+++ opentelemetry-semantic-conventions/src/opentelemetry/semconv/attributes/error_attributes.py
@@ -41,5 +41,5 @@ Note: The `error.type` SHOULD be predictable, and SHOULD have low cardinality.
 
 
 class ErrorTypeValues(Enum):
-    OTHER: Final = "_OTHER"
+    OTHER = "_OTHER"
     """A fallback error value to be used when the instrumentation doesn't define a custom value."""