+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 - Added py.typed file to top-level module ([#4084](https://github.com/open-telemetry/opentelemetry-python/pull/4084))
+- Drop Final annotation from Enum in semantic conventions
+  ([#4085](https://github.com/open-telemetry/opentelemetry-python/pull/4085))
 
 ## Version 1.26.0/0.47b0 (2024-07-25)